SOCINT-57 split out response code into ResponseBuilder
