+++ src/main/java/uk/gov/ons/ctp/integration/mock/data/DataRepository.java
@@ -2,7 +2,6 @@ package uk.gov.ons.ctp.integration.mock.data;
 
 import static uk.gov.ons.ctp.common.log.ScopedStructuredArguments.kv;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
@@ -24,21 +23,13 @@ import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.integration.mock.Constants;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexPartialAddressDTO;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexPartialResultsDTO;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexPostcodeAddressDTO;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexPostcodeResultsDTO;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexRhPostcodeAddressDTO;
-import uk.gov.ons.ctp.integration.mock.ai.model.AddressIndexRhPostcodeResultsDTO;
 import uk.gov.ons.ctp.integration.mock.endpoint.RequestType;
 
 @Slf4j
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class DataRepository {
+public final class DataRepository {
 
   public static String read(RequestType requestType, String name) throws IOException, CTPException {
     String baseFileName = normaliseFileName(name);
@@ -153,98 +144,5 @@ public class DataRepository {
   private static String denormaliseFileName(String name) {
     String baseName = name.replace(".json", "");
     return baseName.replaceAll("-", " ");
-  }
-
-  @SuppressWarnings("unchecked")
-  public static ResponseEntity<Object> simulateResponse(
-      RequestType requestType, String name, int offset, int limit)
-      throws IOException, CTPException {
-
-    HttpStatus responseStatus = HttpStatus.OK;
-    Object response = null;
-    String responseText = DataRepository.read(requestType, name);
-
-    if (responseText != null) {
-      // Convert captured AI response to an object, and return the target subset of data
-      if (requestType.getResponseClass().equals(String.class)) {
-        // Don't push it through Jackson, as it is already in String format
-        response = responseText;
-      } else {
-        response =
-            new ObjectMapper().readerFor(requestType.getResponseClass()).readValue(responseText);
-      }
-
-      switch (requestType) {
-        case AI_RH_POSTCODE:
-          AddressIndexRhPostcodeResultsDTO rhPostcodes =
-              (AddressIndexRhPostcodeResultsDTO) response;
-          List<AddressIndexRhPostcodeAddressDTO> rhPostcodeAddresses =
-              (List<AddressIndexRhPostcodeAddressDTO>)
-                  subset(rhPostcodes.getResponse().getAddresses(), offset, limit);
-          rhPostcodes.getResponse().setAddresses(rhPostcodeAddresses);
-          rhPostcodes.getResponse().setOffset(offset);
-          rhPostcodes.getResponse().setLimit(limit);
-          // Replicate the counting down of the confidence score
-          int confidence = 100000 + rhPostcodeAddresses.size();
-          for (AddressIndexRhPostcodeAddressDTO address : rhPostcodeAddresses) {
-            address.setConfidenceScore(confidence--);
-          }
-          break;
-        case AI_PARTIAL:
-          AddressIndexPartialResultsDTO partial = (AddressIndexPartialResultsDTO) response;
-          List<AddressIndexPartialAddressDTO> partialAddresses =
-              (List<AddressIndexPartialAddressDTO>)
-                  subset(partial.getResponse().getAddresses(), offset, limit);
-          partial.getResponse().setAddresses(partialAddresses);
-          partial.getResponse().setOffset(offset);
-          partial.getResponse().setLimit(limit);
-          break;
-        case AI_POSTCODE:
-          AddressIndexPostcodeResultsDTO postcodes = (AddressIndexPostcodeResultsDTO) response;
-          List<AddressIndexPostcodeAddressDTO> postcodeAddresses =
-              (List<AddressIndexPostcodeAddressDTO>)
-                  subset(postcodes.getResponse().getAddresses(), offset, limit);
-          postcodes.getResponse().setAddresses(postcodeAddresses);
-          postcodes.getResponse().setOffset(offset);
-          postcodes.getResponse().setLimit(limit);
-        case AI_EQ:
-        case AI_RH_UPRN:
-        case CASE_UPRN:
-        case CASE_ID:
-        case CASE_QID:
-        case CASE_REF:
-          // Nothing to do these types
-          break;
-        default:
-          throw new CTPException(
-              CTPException.Fault.SYSTEM_ERROR, "Unrecognised request type: " + requestType.name());
-      }
-    } else {
-      // 404 - not found
-      responseStatus = requestType.getNotFoundHttpStatus();
-      if (requestType.isAddressType()) {
-        responseText = DataRepository.read(requestType, Constants.NO_DATA_FILE_NAME);
-      } else {
-        responseText = null;
-      }
-      if (responseText == null) {
-        responseText = "Data not found";
-      } else {
-        // Customise the not-found response by replacing any place holders with actual values
-        String placeholderName = requestType.getPlaceholderName();
-        String fullPlaceholderName = "%" + placeholderName + "%";
-        responseText = responseText.replace(fullPlaceholderName, name);
-      }
-      response = responseText;
-    }
-    return new ResponseEntity<Object>(response, responseStatus);
-  }
-
-  private static List<?> subset(List<?> addresses, int offset, int limit) {
-    if (offset > addresses.size() || offset < 0) {
-      return new ArrayList<Object>();
-    }
-    int toIndex = Math.min(offset + limit, addresses.size());
-    return addresses.subList(offset, toIndex);
   }
 }
